Add support for top crypto exchanges
Refactor into module
